#3 Change Return Type_Divide Function

+++ src/main/java/edu/handong/csee/pp1/bigdata/lab1_1/Calculator.java
@@ -99,7 +99,7 @@ public class Calculator {
 		return first * second;
 	}
 
-	int divide(int first, int second) {
+	float divide(int first, int second) {
 		countForAnyCompution++;
 		localCount++;
 		return first / second;
